pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "akili"
version = "0.1.0"
description = "The Reasoning Control Plane for Mission-Critical Engineering — deterministic verification for technical documentation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Akili" }]
keywords = ["verification", "documentation", "gemini", "coordinate-grounding", "deterministic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "fastapi>=0.100,<1.0",
    "uvicorn[standard]>=0.22,<1.0",
    "python-multipart>=0.0.6,<1.0",
    "google-generativeai>=0.3,<2.0",
    "pymupdf>=1.23,<2.0",
    "pdf2image>=1.16,<2.0",
    "slowapi>=0.1.9,<1.0",
    "psycopg2-binary>=2.9,<3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7,<9", "pytest-cov>=4,<6", "pytest-asyncio>=0.21,<1", "ruff>=0.1,<1", "flake8>=6,<8", "flake8-pyproject>=1,<2"]
auth = ["firebase-admin>=6.0,<7.0"]
verify = ["z3-solver>=4.12,<5.0"]

[project.scripts]
akili-serve = "akili.api.serve:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", ".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-m 'not integration'"
markers = [
    "integration: live Gemini API tests (require GOOGLE_API_KEY)",
]
